+++ src/main/kotlin/com/wafflestudio/csereal/core/academics/api/v1/AcademicsController.kt
@@ -3,7 +3,6 @@ package com.wafflestudio.csereal.core.academics.api.v1
 import com.wafflestudio.csereal.common.aop.AuthenticatedStaff
 import com.wafflestudio.csereal.common.enums.LanguageType
 import com.wafflestudio.csereal.core.academics.api.req.CreateYearReq
-import com.wafflestudio.csereal.core.academics.api.req.UpdateScholarshipPageReq
 import com.wafflestudio.csereal.core.academics.api.req.UpdateSingleReq
 import com.wafflestudio.csereal.core.academics.api.req.UpdateYearReq
 import com.wafflestudio.csereal.core.academics.dto.*
@@ -87,7 +86,7 @@ class AcademicsController(
     ) = academicsService.deleteAcademicsYearResponse(language, studentType, postType, year)
 
     //교과목 정보
-
+    @Deprecated(message = "Use V2 API")
     @GetMapping("/{studentType}/courses")
     fun readAllCourses(
         @RequestParam(required = false, defaultValue = "ko") language: String,
@@ -111,14 +110,6 @@ class AcademicsController(
         @RequestPart newAttachments: List<MultipartFile>?
     ) = academicsService.updateDegreeRequirements(language, request, newAttachments)
 
-    @AuthenticatedStaff
-    @PutMapping("/{studentType}/scholarship")
-    fun updateScholarshipPage(
-        @RequestParam(required = false, defaultValue = "ko") language: String,
-        @PathVariable studentType: String,
-        @RequestBody request: UpdateScholarshipPageReq
-    ) = academicsService.updateScholarshipPage(language, studentType, request)
-
     @GetMapping("/{studentType}/scholarship")
     fun readAllScholarship(
         @RequestParam(required = false, defaultValue = "ko") language: String,
@@ -127,6 +118,7 @@ class AcademicsController(
         return ResponseEntity.ok(academicsService.readAllScholarship(language, studentType))
     }
 
+    @Deprecated(message = "Use V2 API")
     @GetMapping("/scholarship/{scholarshipId}")
     fun getScholarship(
         @PathVariable scholarshipId: Long
